Migrate YamlSecurityTest
