+++ its/plugin/fast-tests/src/test/java/com/sonar/javascript/it/plugin/TestUtils.java
@@ -133,4 +133,8 @@ public class TestUtils {
       throw new UncheckedIOException(e);
     }
   }
+
+  public static boolean isUcfgFile(Path path, BasicFileAttributes attrs) {
+    return attrs.isRegularFile() && path.getFileName().toString().endsWith(".ucfgs");
+  }
 }
